requirements.txt
requests~=2.32.3
pandas~=2.2.3
yfinance~=0.2.52
python-dotenv~=1.0.1
python-telegram-bot~=20.8